Move project from yarn to npm
We've used `deyarn` to move this template project from `yarn` to `npm`. This decision was made because we experienced participants encountering some unnecessary friction in getting npm and yarn set up on their machines and we'd like to avoid as much of those troubles as possible going forward.

To get more details on the conversation, see [this Slack conversation](https://the-collab-lab.slack.com/archives/GLQ89ULE6/p1576808365001200).