Update README with setup instructions
This commit adds instructions on how to setup the project as well as instructions for cohort members to add their names to the CONTRIBUTORS.md file. I think maybe this oughta be more clearly laid out, but it's been a looonnggg day, so maybe tomorrow -- y'all should feel free to contribute and refine this. 😝