Add estimates into lib/ and update calculateEstimate
An updated estimates function was added to lib/ of the shopping
list template to eleminate the need for supplying a gist with the
accompanying issue.

The calculateEstimate function was updated to take into account the
condition when numberOfPurchases is less than 2. For less than 2
the estimate is the most recent interval, while a
weighted average is taken when 2 or more purchases have been made.