Add recap to ORIENTATION doc
This commit adds the idea of a recap to the ORIENTATION doc. I have found that any time I have run orientation, I end up giving a quick, high-level recap because there is just so much new info and it can be a little overwhelming for new Collabies. My idea here is to solidify that recap so that anyone running the end of an Orientation has a solid outline for making the new Collabies feel at ease and better understand all the information they just reviewed during Orientation.